backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
PyJWT
cryptography
pydantic
sqlalchemy-utils
requests